REL-3976: test cases
